Removing depenency of WPane
WPane was from an earlier UI that would zoom into the area where a darworm was selcting a new direction.  This proved to be too distracting however and tended to cause a loss of contect for players.  I am going to leave the (now undeeded) file around for a little while for sample code to draw grid cells